shut Warning about %lu format not matching sOutput.length()
size_t is unsigned int or long unsigned int depending on platform

+++ ce_main_app/webserver/api/resource/floppyresource.cpp
@@ -61,7 +61,7 @@ bool FloppyResource::dispatch(mg_connection *conn, mg_request_info *req_info, st
         }
         stringStream << "}"; 
         std::string sJson=stringStream.str();
-        mg_printf(conn, "Content-Length: %lu\r\n\r\n",sJson.length());   // Always set Content-Length
+        mg_printf(conn, "Content-Length: %lu\r\n\r\n",(unsigned long)sJson.length());   // Always set Content-Length
         mg_write(conn, sJson.c_str(), sJson.length());                  // send content
         return true;
     }
